Make cmdJOIN not the most broken thing ever
The problem with its implementation was the split of channels and keys going
way out of sync. Joining multiple channels in a single command is hard to do
this way because it's impossible to tell what key to use with what channel.
This is solved by removing the functionality to join multiple channels
altogether. It's not that hard to send multiple commands anyway.