Log: Allow slashes in dates
Update log.py

Got rid of the variable. (Much neater, cool idea.)
Not too sure if I got the placement (or wording) of the other
suggestions right. Used the standard error message.

Edit by Heufneutje: Squashed the commits and reworded the commit
message to make it consistent with the other commits.

Closes #39